+++ libs/crypto/src/lib/ciphers-current.ts
@@ -23,7 +23,7 @@ import sodium from 'libsodium-wrappers';
 import * as cc from './cipher.consts';
 
 import {
-   getRandom48,
+   getRandom,
    numToBytes,
    bytesToNum,
    BYOBStreamReader,
@@ -31,20 +31,7 @@ import {
    getArrayBuffer,
    ensureArrayBuffer,
 } from './utils';
-
-
-export type EParams = {
-   readonly alg: string;
-   readonly ic: number;
-   readonly lp: number;
-   readonly lpEnd: number;
-};
-
-
-export type PWDProvider = (
-   cdInfo: CipherDataInfo,
-   encrypting: boolean
-) => Promise<[string, string | undefined]>;
+import { KeyProvider, PWDKeyProvider } from './keys';
 
 export enum CipherState {
    Error,
@@ -65,9 +52,8 @@ export type CipherDataInfo = {
    readonly ic: number;
    readonly lp: number;
    readonly lpEnd: number;
-   readonly iv: Uint8Array;
-   readonly slt: Uint8Array;
-   readonly hint?: string;
+   readonly slt: Uint8Array<ArrayBuffer>;
+   hint?: string | undefined;
 };
 
 
@@ -81,24 +67,13 @@ const KDF_CTX_BLOCK_V7 = "Blck_Key";
 
 export abstract class Ciphers {
 
-   protected _ek?: Uint8Array;
-   protected _sk?: Uint8Array;
-
-   protected _reader: BYOBStreamReader;
-   protected _userCred?: Uint8Array;
    protected _state: CipherState;
 
    protected constructor(
-      userCred: Uint8Array,
-      reader: BYOBStreamReader
+      protected _keyProvider: KeyProvider,
+      protected _reader: BYOBStreamReader
    ) {
-      if (!userCred || userCred.byteLength != cc.USERCRED_BYTES) {
-         throw new Error('Invalid userCred');
-      }
-
       this._state = CipherState.Initialized;
-      this._userCred = userCred;
-      this._reader = reader;
    }
 
    public finishedState() {
@@ -112,18 +87,7 @@ export abstract class Ciphers {
    }
 
    protected _purge() {
-      // Don't overwrite because this is only a reference
-      this._userCred = undefined;
-
-      // overwrite to clear
-      if (this._ek) {
-         crypto.getRandomValues(this._ek);
-         this._ek = undefined;
-      }
-      if (this._sk) {
-         crypto.getRandomValues(this._sk);
-         this._sk = undefined;
-      }
+      this._keyProvider.purge();
       this._reader.cleanup();
    }
 
@@ -135,8 +99,19 @@ export abstract class Ciphers {
       maxMillis: number
    ): Promise<[number, number, number]> {
 
+      const cdInfo = {
+         ver: cc.CURRENT_VERSION,
+         alg: 'AES-GCM',
+         ic: testSize,
+         lp: 1,
+         lpEnd: 1
+      };
+      const keyProvider = new PWDKeyProvider(
+         crypto.getRandomValues(new Uint8Array(32)),
+         ['AVeryBogusPwd', '']
+      );
       const start = Date.now();
-      await _genCipherKey('AES-GCM', testSize, 'AVeryBogusPwd', crypto.getRandomValues(new Uint8Array(32)), new Uint8Array(cc.SLT_BYTES));
+      await keyProvider.getCipherKey(true);
       const test_millis = Date.now() - start;
 
       const hashRate = testSize / test_millis;
@@ -319,40 +294,29 @@ export abstract class Encipher extends Ciphers {
    //protected static readonly READ_SIZE_START = 9;
    //protected static readonly READ_SIZE_MAX = READ_SIZE_START * 16
 
-
    protected constructor(
-      userCred: Uint8Array,
+      keyProvider: KeyProvider,
       reader: BYOBStreamReader
    ) {
-      super(userCred, reader);
+      super(keyProvider, reader);
    }
 
    // The encryptBlock functions return CipherDataBlock instead of
    // an opaque byte array to reduce copying (the caller must write
    // all the blocks in order)
-   async encryptBlock(
-      eparams: EParams,
-      pwdProvider: PWDProvider
-   ): Promise<CipherDataBlock> {
+   async encryptBlock(): Promise<CipherDataBlock> {
 
       if (this._state == CipherState.Initialized) {
-         return this.encryptBlock0(eparams, pwdProvider);
+         return this.encryptBlock0();
       } else if (this._state == CipherState.Block0Done) {
-         return this.encryptBlockN(eparams);
+         return this.encryptBlockN();
       } else {
          throw new Error(`Encipher invalid state ${this._state}`);
       }
    }
 
-   public abstract encryptBlock0(
-      eparams: EParams,
-      pwdProvider: PWDProvider
-   ): Promise<CipherDataBlock>;
-
-   public abstract encryptBlockN(
-      eparams: EParams
-   ): Promise<CipherDataBlock>;
-
+   public abstract encryptBlock0(): Promise<CipherDataBlock>;
+   public abstract encryptBlockN(): Promise<CipherDataBlock>;
 };
 
 
@@ -409,13 +373,12 @@ export class EncipherV7 extends Encipher {
    private _blockNum;
    private _readTarget;
    private _lastMac?: Uint8Array;
-   private _slt?: Uint8Array; // stored as class member to help with testings
 
    constructor(
-      userCred: Uint8Array,
+      keyProvider: KeyProvider,
       reader: BYOBStreamReader
    ) {
-      super(userCred, reader);
+      super(keyProvider, reader);
 
       // Assign in ctor to allow for monkey patching in tests
       this._readTarget = Encipher.READ_SIZE_START;
@@ -424,10 +387,6 @@ export class EncipherV7 extends Encipher {
    }
 
    protected override _purge() {
-      if (this._slt) {
-         crypto.getRandomValues(this._slt);
-         this._slt = undefined;
-      }
       if (this._lastMac) {
          crypto.getRandomValues(this._lastMac);
          this._lastMac = undefined;
@@ -449,22 +408,11 @@ export class EncipherV7 extends Encipher {
    // 5. Sign addtional data + cipher text with signing key
    // 6. Return all the parts of the cipherdata
    //
-   override async encryptBlock0(
-      eparams: EParams,
-      pwdProvider: PWDProvider
-   ): Promise<CipherDataBlock> {
-
-      let hk: Uint8Array | undefined;
-      let hIV: Uint8Array | undefined;
+   override async encryptBlock0(): Promise<CipherDataBlock> {
 
       try {
-         EncipherV7.validateEParams(eparams);
          if (this._state != CipherState.Initialized) {
             throw new Error(`Encipher invalid state ${this._state}`);
-         }
-
-         if (this._sk || this._ek) {
-            throw new Error('encryptBlock0 should only be called once');
          }
 
          const [clearBuffer, done] = await this._reader.readAvailable(
@@ -486,47 +434,21 @@ export class EncipherV7 extends Encipher {
             };
          }
 
-         // Create a new salt and IV each time a key is derviced from the password.
-         // https://crypto.stackexchange.com/questions/53032/salt-for-non-stored-passwords
-         const randomArray = getRandom48();
-
-         const ivBytes = Number(cc.AlgInfo[eparams.alg]['iv_bytes']);
-         this._slt = randomArray.slice(0, cc.SLT_BYTES);
-         const iv = randomArray.slice(cc.SLT_BYTES, cc.SLT_BYTES + ivBytes);
-
-         const [pwd, hint] = await pwdProvider({
-               ver: cc.VERSION7,
-               alg: eparams.alg,
-               ic: eparams.ic,
-               slt: this._slt,
-               lp: eparams.lp,
-               lpEnd: eparams.lpEnd,
-               iv: iv,
-               hint: undefined
-            },
-            true
-         );
-         if (hint && hint.length > cc.HINT_MAX_LEN) {
-            throw new Error('Hint length exceeds: ' + cc.HINT_MAX_LEN);
-         }
-         if (!pwd) {
-            throw new Error('Missing password');
-         }
-
-         this._sk = _genSigningKey(this._userCred!, this._slt);
-         this._ek = await _genCipherKey(eparams.alg, eparams.ic, pwd, this._userCred!, this._slt);
+         const cdInfo = this._keyProvider.getCipherDataInfo();
+         const iv = getRandom(Number(cc.AlgInfo[cdInfo.alg]['iv_bytes']));
+         const ek = await this._keyProvider.getCipherKey(true);
 
          let encryptedHint: Uint8Array<ArrayBufferLike> = new Uint8Array(0);
-         if (hint) {
+         if (cdInfo.hint) {
             const maxHintBytes = cc.ENCRYPTED_HINT_MAX_BYTES - cc.AUTH_TAG_MAX_BYTES;
-            let hintBytes = bytesFromString(hint, maxHintBytes);
+            let hintBytes = bytesFromString(cdInfo.hint, maxHintBytes);
 
             // It's possible that even a single character (e.g. emoji) might exceed maxHintBytes
             // If so proceed without a hint.
             if (hintBytes.byteLength > 0) {
-               [hk, hIV] = _genHintCipherKeyAndIV(this._userCred!, iv, this._slt);
+               const [hk, hIV] = await this._keyProvider.getHintCipherKeyAndIV(iv);
                encryptedHint = await EncipherV7._doEncrypt(
-                  eparams.alg,
+                  cdInfo.alg,
                   hk,
                   hIV,
                   hintBytes
@@ -535,26 +457,26 @@ export class EncipherV7 extends Encipher {
          }
 
          const additionalData = Ciphers._encodeAdditionalData({
-            alg: eparams.alg,
-            iv,
+            alg: cdInfo.alg,
+            iv: iv,
             term: done,
-            ic: eparams.ic,
-            lp: eparams.lp,
-            lpEnd: eparams.lpEnd,
-            slt: this._slt,
+            ic: cdInfo.ic,
+            lp: cdInfo.lp,
+            lpEnd: cdInfo.lpEnd,
+            slt: cdInfo.slt,
             encryptedHint
          });
 
          // Only block0 uses the root cipher key. Simplifies backward compat and is no less secure
          const encryptedData = await EncipherV7._doEncrypt(
-            eparams.alg,
-            this._ek,
+            cdInfo.alg,
+            ek,
             iv,
             clearBuffer,
             additionalData,
          );
 
-         const headerData = this._createHeader(
+         const headerData = await this._createHeader(
             encryptedData,
             additionalData
          );
@@ -577,28 +499,14 @@ export class EncipherV7 extends Encipher {
          this.errorState();
          console.error(err);
          throw err;
-      } finally {
-         if (hk) {
-            crypto.getRandomValues(hk);
-            hk = undefined;
-         }
       }
    }
 
-   override async encryptBlockN(
-      eparams: EParams
-   ): Promise<CipherDataBlock> {
-
-      let bk: Uint8Array | undefined;
+   override async encryptBlockN(): Promise<CipherDataBlock> {
 
       try {
-         EncipherV7.validateEParams(eparams);
          if (this._state != CipherState.Block0Done) {
             throw new Error(`Encipher invalid state ${this._state}`);
-         }
-
-         if (!this._sk || !this._ek) {
-            throw new Error('Data not initialized, encrypt block0 first');
          }
 
          this._readTarget = Math.min(this._readTarget * 2, Encipher.READ_SIZE_MAX);
@@ -619,29 +527,26 @@ export class EncipherV7 extends Encipher {
             }
          }
 
-         const randomArray = getRandom48();
-
-         const ivBytes = Number(cc.AlgInfo[eparams.alg]['iv_bytes']);
-         const iv = randomArray.slice(0, ivBytes);
+         const cdInfo = this._keyProvider.getCipherDataInfo();
+         const iv = getRandom(Number(cc.AlgInfo[cdInfo.alg]['iv_bytes']));
+         const bk = await this._keyProvider.getBlockCipherKey(this._blockNum);
+         this._blockNum += 1;
 
          const additionalData = Ciphers._encodeAdditionalData({
-            alg: eparams.alg,
-            iv: iv,
+            alg: cdInfo.alg,
+            iv,
             term: done,
          });
 
-         bk = _genBlockCipherKey(this._ek, this._slt!, this._blockNum);
-         this._blockNum += 1;
-
          const encryptedData = await EncipherV7._doEncrypt(
-            eparams.alg,
+            cdInfo.alg,
             bk,
             iv,
             clearBuffer,
             additionalData,
          );
 
-         const headerData = this._createHeader(
+         const headerData = await this._createHeader(
             encryptedData,
             additionalData
          );
@@ -662,33 +567,6 @@ export class EncipherV7 extends Encipher {
          this.errorState();
          console.error(err);
          throw err;
-      } finally {
-         if (bk) {
-            crypto.getRandomValues(bk);
-            bk = undefined;
-         }
-      }
-   }
-
-   public static validateEParams(eparams: EParams) {
-      const {
-         alg,
-         ic,
-         lp,
-         lpEnd
-      } = eparams;
-
-      if (!Ciphers.validateAlg(alg)) {
-         throw new Error('Invalid alg type of: ' + alg);
-      }
-      if (ic < cc.ICOUNT_MIN || ic > cc.ICOUNT_MAX) {
-         throw new Error('Invalid ic of: ' + ic);
-      }
-      if (lpEnd < 1 || lpEnd > cc.LP_MAX) {
-         throw new Error('Invalid lpEnd: ' + lpEnd);
-      }
-      if (lp < 1 || lp > lpEnd) {
-         throw new Error('Invalid lp: ' + lp);
       }
    }
 
@@ -761,14 +639,13 @@ export class EncipherV7 extends Encipher {
       return encryptedBytes;
    }
 
-
    // Seperated out and made public for testing, normal callers should not use this function
-   public _createHeader(
+   public async _createHeader(
       encryptedData: Uint8Array,
       additionalData: Uint8Array
-   ): Uint8Array {
+   ): Promise<Uint8Array> {
 
-      if (!this._lastMac || !this._sk || !encryptedData || !additionalData) {
+      if (!this._lastMac || !encryptedData || !additionalData) {
          throw new Error('Data not initialized');
       }
 
@@ -778,7 +655,8 @@ export class EncipherV7 extends Encipher {
       packer.ver = cc.VERSION7;
       packer.size = payloadBytes;
 
-      const state = sodium.crypto_generichash_init(this._sk, cc.MAC_BYTES);
+      const sk = await this._keyProvider.getSigningKey();
+      const state = sodium.crypto_generichash_init(sk, cc.MAC_BYTES);
       sodium.crypto_generichash_update(state, new Uint8Array(packer.buffer, cc.MAC_BYTES));
       sodium.crypto_generichash_update(state, additionalData);
       sodium.crypto_generichash_update(state, encryptedData);
@@ -809,32 +687,17 @@ export abstract class Decipher extends Ciphers {
 
    protected _blockNum;
    protected _blockData?: BlockData;
-   protected _slt?: Uint8Array<ArrayBuffer>;
-   protected _ic?: number;
-   protected _lp: number;
-   protected _lpEnd: number;
-   protected _hint?: Uint8Array;
 
    protected constructor(
-      userCred: Uint8Array,
+      keyProvider: KeyProvider,
       reader: BYOBStreamReader
    ) {
-      super(userCred, reader);
+      super(keyProvider, reader);
 
       this._blockNum = 1;
-      this._lp = 1; // Not all version support loop # so provide a default
-      this._lpEnd = 1; // Not all version support loop # so provide a default
    }
 
    protected override _purge() {
-      if (this._slt) {
-         crypto.getRandomValues(this._slt);
-         this._slt = undefined;
-      }
-      if (this._hint) {
-         crypto.getRandomValues(this._hint);
-         this._hint = undefined;
-      }
       this._blockData = undefined;
       super._purge();
    }
@@ -845,12 +708,9 @@ export abstract class Decipher extends Ciphers {
    // empty when done)
 
    // Helper that calls Block0 until ~success then BlockN
-   async decryptBlock(
-      pwdProvider: PWDProvider
-   ): Promise<Uint8Array> {
-
+   async decryptBlock(): Promise<Uint8Array> {
       if ([CipherState.Initialized, CipherState.Block0Decoded].includes(this._state)) {
-         return this.decryptBlock0(pwdProvider);
+         return this.decryptBlock0();
       } else if (this._state == CipherState.Block0Done) {
          return this.decryptBlockN();
       } else {
@@ -871,7 +731,6 @@ export abstract class Decipher extends Ciphers {
    // 6. Return cleat text bytes
    //
    public async decryptBlock0(
-      pwdProvider: PWDProvider
    ): Promise<Uint8Array> {
 
       try {
@@ -879,44 +738,18 @@ export abstract class Decipher extends Ciphers {
             throw new Error(`Decipher invalid state ${this._state}`);
          }
 
-         if (this._ek) {
-            throw new Error('Decipher unexpected encryption key');
-         }
-
-         if (this._userCred!.byteLength != cc.USERCRED_BYTES) {
-            throw new Error('Invalid userCred length of: ' + this._userCred!.byteLength);
-         }
-
          // This does MAC check
          await this._decodeBlock0();
-
-         if (!this._blockData || !this._blockData.alg || !this._blockData.iv || !this._blockData.encryptedData || !this._slt || !this._ic) {
+         if (!this._blockData || !this._blockData.alg || !this._blockData.iv || !this._blockData.encryptedData) {
             throw new Error('Data not initialized');
          }
 
-         const [pwd] = await pwdProvider({
-               ver: this._blockData.ver,
-               alg: this._blockData.alg,
-               ic: this._ic,
-               slt: this._slt,
-               lp: this._lp,
-               lpEnd: this._lpEnd,
-               iv: this._blockData.iv,
-               hint: new TextDecoder().decode(this._hint)
-            },
-            false
-         );
-
-         if (!pwd) {
-            throw new Error('password is empty');
-         }
-
-         this._ek = await _genCipherKey(this._blockData.alg, this._ic, pwd, this._userCred!, this._slt);
+         const ek = await this._keyProvider.getCipherKey(false);
 
          // Only block0 uses the root cipher key. Simplifies backward compat and is no less secure
          const decrypted = await Decipher._doDecrypt(
             this._blockData.alg,
-            this._ek,
+            ek,
             this._blockData.iv,
             this._blockData.encryptedData,
             this._blockData.additionalData,
@@ -935,28 +768,25 @@ export abstract class Decipher extends Ciphers {
 
    //public only for testing
    abstract _decodeBlock0(): Promise<void>;
-
    public abstract decryptBlockN(): Promise<Uint8Array>;
 
-   public async getCipherDataInfo(
-   ): Promise<CipherDataInfo> {
-
+   public async getCipherDataInfo(): Promise<CipherDataInfo> {
       // This does MAC check
       await this._decodeBlock0();
-
-      if (!this._blockData || !this._blockData.alg || !this._blockData.iv || !this._slt || !this._ic) {
+      if (!this._blockData || !this._blockData.alg || !this._blockData.iv) {
          throw new Error('Data not initialized');
       }
 
+      const cdInfo = this._keyProvider.getCipherDataInfo();
       return {
          ver: this._blockData.ver,
          alg: this._blockData.alg,
-         ic: this._ic,
-         slt: this._slt,
-         lp: this._lp,
-         lpEnd: this._lpEnd,
-         iv: this._blockData.iv,
-         hint: new TextDecoder().decode(this._hint)
+         ic: cdInfo.ic,
+         slt: cdInfo.slt,
+         lp: cdInfo.lp,
+         lpEnd: cdInfo.lpEnd,
+         // iv: this._blockData.iv,
+         hint: cdInfo.hint
       };
    }
 
@@ -1089,11 +919,11 @@ export class DecipherV67 extends Decipher {
    private _lastFlags: number;
 
    constructor(
-      userCred: Uint8Array,
+      keyProvider: KeyProvider,
       reader: BYOBStreamReader,
       header?: Uint8Array
    ) {
-      super(userCred, reader);
+      super(keyProvider, reader);
       this._header = header;
 
       this._lastMac = new Uint8Array([0]);
@@ -1113,7 +943,6 @@ export class DecipherV67 extends Decipher {
    }
 
    private async _decodeHeader(header?: Uint8Array): Promise<boolean> {
-
       // Need to treat all values an UNTRUSTED since the signature has not yet been
       // validated.
       if (!header) {
@@ -1154,9 +983,6 @@ export class DecipherV67 extends Decipher {
    // with zero length encrypted text since that is not needed
    override async _decodeBlock0(): Promise<void> {
 
-      let hk: Uint8Array | undefined;
-      let hIV: Uint8Array | undefined;
-
       try {
          if (![CipherState.Initialized, CipherState.Block0Decoded].includes(this._state)) {
             throw new Error(`Decipher invalid state ${this._state}`);
@@ -1167,12 +993,8 @@ export class DecipherV67 extends Decipher {
             return;
          }
 
-         if (this._sk) {
-            throw new Error('Decipher unexpected signing key');
-         }
-
+         // This setups up _blockData
          await this._decodeHeader(this._header);
-
          if (!this._blockData) {
             throw new Error('Data not initialized');
          }
@@ -1195,39 +1017,48 @@ export class DecipherV67 extends Decipher {
          this._blockData.flags = extractor.flags;
          this._blockData.alg = extractor.alg;
          this._blockData.iv = extractor.iv;
-         this._slt = extractor.slt;
-         this._ic = extractor.ic;
-         [this._lp, this._lpEnd] = extractor.lpp();
+         const slt = extractor.slt;
+         const ic = extractor.ic;
+         const [lp, lpEnd] = extractor.lpp();
          const encryptedHint = extractor.hint;
          this._blockData.encryptedData = extractor.remainder('edata');
 
-         // V4 additional data is the payload minus encrypted data
+         // Since V4, additional data is the payload minus encrypted data
          this._blockData.additionalData = new Uint8Array(
             payload.buffer,
             payload.byteOffset,
             extractor.offset - this._blockData.encryptedData.byteLength
          );
 
-         this._sk = _genSigningKey(this._userCred!, this._slt, this._blockData.ver);
+         const cdInfo: CipherDataInfo = {
+            ver: this._blockData.ver,
+            alg: this._blockData.alg,
+            ic: ic,
+            lp: lp,
+            lpEnd: lpEnd,
+            slt: slt
+         };
+         this._keyProvider.setCipherDataInfo(cdInfo);
 
          // Avoiding the Doom Principle and verify signature before crypto operations.
          // Aka, check MAC as soon as possible after we have the signing key and data.
          // Might be cleaner to do this elsewhere, but keeping it at the lowest level
          // ensures we don't skip the step
-         const validMac: boolean = this._verifyMAC();
+         const validMac: boolean = await this._verifyMAC();
          if (!validMac) {
             throw new Error('Invalid MAC error');
          }
 
-         let hint: Uint8Array<ArrayBufferLike> = new Uint8Array(0);
          if (encryptedHint!.byteLength != 0) {
-            [hk, hIV] = _genHintCipherKeyAndIV(this._userCred!, this._blockData.iv, this._slt, this._blockData.ver);
-            this._hint = await Decipher._doDecrypt(
+            const [hk, hIV] = await this._keyProvider.getHintCipherKeyAndIV(this._blockData.iv);
+            const hintBytes = await Decipher._doDecrypt(
                this._blockData.alg,
                hk,
                hIV,
                encryptedHint
             );
+            cdInfo.hint = new TextDecoder().decode(hintBytes);
+            this._keyProvider.setCipherDataInfo(cdInfo);
          }
 
          this._state = CipherState.Block0Decoded;
@@ -1239,25 +1070,15 @@ export class DecipherV67 extends Decipher {
          throw err;
       } finally {
          this._header = undefined;
-         if (hk) {
-            crypto.getRandomValues(hk);
-            hk = undefined;
-         }
       }
    }
 
    public override async decryptBlockN(
    ): Promise<Uint8Array> {
 
-      let bk: Uint8Array | undefined;
-
       try {
          if (this._state != CipherState.Block0Done) {
             throw new Error(`Decipher invalid state ${this._state}`);
-         }
-
-         if (!this._sk || !this._ek) {
-            throw new Error('Data not initialized, decrypt block0 first');
          }
 
          // This does MAC check
@@ -1268,11 +1089,11 @@ export class DecipherV67 extends Decipher {
             return new Uint8Array(0);
          }
 
-         if (!this._blockData || !this._blockData.alg || !this._ek || !this._blockData.iv || !this._blockData.encryptedData) {
+         if (!this._blockData || !this._blockData.alg || !this._blockData.iv || !this._blockData.encryptedData) {
             throw new Error('Data not initialized');
          }
 
-         bk = _genBlockCipherKey(this._ek, this._slt!, this._blockNum, this._blockData.ver);
+         const bk = await this._keyProvider.getBlockCipherKey(this._blockNum);
          this._blockNum += 1;
 
          const decrypted = await Decipher._doDecrypt(
@@ -1290,10 +1111,6 @@ export class DecipherV67 extends Decipher {
          throw err;
       } finally {
          this._blockData = undefined;
-         if (bk) {
-            crypto.getRandomValues(bk);
-            bk = undefined;
-         }
       }
    }
 
@@ -1305,6 +1122,7 @@ export class DecipherV67 extends Decipher {
       // Need to treat all values an UNTRUSTED since the signature has not yet been
       // validated, Extractor does test each value for valid ranges as we unpack
       try {
+         // This setups up _blockData
          const done = await this._decodeHeader();
          if (done) {
             if (this._lastFlags !== 1) {
@@ -1319,7 +1137,6 @@ export class DecipherV67 extends Decipher {
          if (this._lastFlags === 1) {
             throw new Error(`Extra data block ${this._state}`);
          }
-
          if (!this._blockData) {
             throw new Error('Data not initialized');
          }
@@ -1347,11 +1164,18 @@ export class DecipherV67 extends Decipher {
             extractor.offset - this._blockData.encryptedData.byteLength
          );
 
+         // Since V4, additional data is the payload minus encrypted data
+         this._blockData.additionalData = new Uint8Array(
+            payload.buffer,
+            payload.byteOffset,
+            extractor.offset - this._blockData.encryptedData.byteLength
+         );
+
          // Avoiding the Doom Principle and verify signature before crypto operations.
          // Aka, check MAC as soon as possible after we  have the signing key and data.
          // Might be cleaner to do this elswhere, but keeping it at the lowest level
          // ensures we don't skip the step
-         const validMac: boolean = this._verifyMAC();
+         const validMac: boolean = await this._verifyMAC();
          if (!validMac) {
             throw new Error('Invalid MAC error');
          }
@@ -1365,10 +1189,10 @@ export class DecipherV67 extends Decipher {
       }
    }
 
-   private _verifyMAC(): boolean {
+   private async _verifyMAC(): Promise<boolean> {
 
       if (!this._blockData || !this._blockData.payloadSize || !this._blockData.ver || !this._blockData.additionalData ||
-         !this._sk || !this._blockData.encryptedData || !this._blockData.mac || !this._lastMac) {
+        !this._blockData.encryptedData || !this._blockData.mac || !this._lastMac) {
          throw new Error('Data not initialized');
       }
 
@@ -1379,7 +1203,8 @@ export class DecipherV67 extends Decipher {
       headerPortion.set(encVerBytes);
       headerPortion.set(encSizeBytes, cc.VER_BYTES);
 
-      const state = sodium.crypto_generichash_init(this._sk, cc.MAC_BYTES);
+      const sk = await this._keyProvider.getSigningKey();
+      const state = sodium.crypto_generichash_init(sk, cc.MAC_BYTES);
 
       sodium.crypto_generichash_update(state, headerPortion);
       sodium.crypto_generichash_update(state, this._blockData.additionalData);
@@ -1396,7 +1221,6 @@ export class DecipherV67 extends Decipher {
 
       throw new Error('Invalid MAC signature');
    }
-
 }
 
 
@@ -1670,139 +1494,3 @@ export class Packer {
       this.pack('size', numToBytes(payloadSize, cc.PAYLOAD_SIZE_BYTES));
    }
 }
-
-
-// Exported for testing, normal callers should not need this
-export async function _genCipherKey(
-   alg: string,
-   ic: number,
-   pwd: string,
-   userCred: Uint8Array,
-   slt: Uint8Array
-): Promise<Uint8Array<ArrayBuffer>> {
-
-   if (!Ciphers.validateAlg(alg)) {
-      throw new Error('Invalid alg type of: ' + alg);
-   }
-   if (ic < cc.ICOUNT_MIN || ic > cc.ICOUNT_MAX) {
-      throw new Error('Invalid ic of: ' + ic);
-   }
-   if (!pwd) {
-      throw new Error('Invalid empty password');
-   }
-   if (userCred.byteLength != cc.USERCRED_BYTES) {
-      throw new Error("Invalid userCred length of: " + userCred.byteLength);
-   }
-   if (slt.byteLength != cc.SLT_BYTES) {
-      throw new Error("Invalid slt length of: " + slt.byteLength);
-   }
-
-   const pwdBytes = new TextEncoder().encode(pwd);
-   const rawMaterial = new Uint8Array(pwdBytes.byteLength + cc.USERCRED_BYTES)
-   rawMaterial.set(pwdBytes);
-   rawMaterial.set(userCred, pwdBytes.byteLength);
-
-   const ekMaterial = await crypto.subtle.importKey(
-      'raw',
-      rawMaterial,
-      'PBKDF2',
-      false,
-      ['deriveBits', 'deriveKey']
-   );
-
-   // overwrite to clear userCred and pwd
-   crypto.getRandomValues(rawMaterial);
-   crypto.getRandomValues(pwdBytes);
-
-   // A bit of a hack, but subtle doesn't support other algorithms... so lie. This
-   // is safe because the key is exported as bits and used in libsodium when not
-   // AES-GCM. TODO: If more non-browser cipher are added, make this more generic.
-   const useAlg = 'AES-GCM';
-
-   let subtleKey: CryptoKey | undefined = await crypto.subtle.deriveKey(
-      {
-         name: 'PBKDF2',
-         salt: getArrayBuffer(slt),
-         iterations: ic,
-         hash: 'SHA-512',
-      },
-      ekMaterial,
-      { name: useAlg, length: 256 },
-      true,
-      ['encrypt', 'decrypt']
-   );
-
-   const exported = await crypto.subtle.exportKey("raw", subtleKey);
-   subtleKey = undefined;
-   return new Uint8Array(exported);
-}
-
-// Exported for testing, normal callers should not need this
-export function _genBlockCipherKey(
-   cipherKey: Uint8Array,
-   slt: Uint8Array,
-   blockNum: number,
-   ver: number = cc.CURRENT_VERSION
-): Uint8Array<ArrayBuffer> {
-   const ctx = ver === cc.VERSION6 ? KDF_CTX_BLOCK_V6 : KDF_CTX_BLOCK_V7;
-   return _genDerivedKey(cipherKey, slt, ctx, blockNum, ver);
-}
-
-
-export function _genDerivedKey(
-   master: Uint8Array,
-   slt: Uint8Array,
-   purpose: string,
-   instance: number,
-   ver: number
-): Uint8Array<ArrayBuffer> {
-
-   let mixedKey: Uint8Array;
-
-   // VERSION7 adds a salt to key derivations
-   if (ver === cc.VERSION7) {
-      if (purpose.length != 8) {
-         throw new Error('Invalid purpose length: ' + purpose.length);
-      }
-
-      // because crypto_kdf_derive_from_key does not take a salt, we first merge salt and
-      // master into a single hash.
-      mixedKey = sodium.crypto_generichash(cc.KEY_BYTES, slt, master);
-   } else {
-      mixedKey = master;
-   }
-
-   return ensureArrayBuffer(sodium.crypto_kdf_derive_from_key(
-      master.byteLength,
-      instance,
-      purpose.slice(0, 8),
-      mixedKey
-   ));
-}
-
-
-// Exported for testing and old deciphers, normal callers should not need this
-export function _genHintCipherKeyAndIV(
-   userCred: Uint8Array,
-   iv: Uint8Array<ArrayBuffer>,
-   slt: Uint8Array,
-   ver: number = cc.CURRENT_VERSION
-): [Uint8Array<ArrayBuffer>, Uint8Array<ArrayBuffer>] {
-   const ctx = ver === cc.VERSION6 ? KDF_CTX_HINT_V6 : KDF_CTX_HINT_V7;
-   return [
-      _genDerivedKey(userCred, slt, ctx, 1, ver),
-      ver === cc.VERSION6 ? iv : _genDerivedKey(iv, slt, ctx, 2, ver)
-   ];
-}
-
-
-// Exported for testing, normal callers should not need this
-export function _genSigningKey(
-   userCred: Uint8Array,
-   slt: Uint8Array,
-   ver: number = cc.CURRENT_VERSION
-): Uint8Array<ArrayBuffer> {
-   const ctx = ver === cc.VERSION6 ? KDF_CTX_SIGNING_V6 : KDF_CTX_SIGNING_V7;
-   return _genDerivedKey(userCred, slt, ctx, 1, ver);
-}
-