+++ libs/crypto/src/lib/utils.ts
@@ -32,7 +32,6 @@ export function ensureArrayBuffer(value: Uint8Array): Uint8Array<ArrayBuffer> {
    if (hasArrayBuffer(value)) {
       return value;
    } else {
-      //@ts-ignore (something in angular tool-chain seems to be using old types)
       return value.slice(0);
    }
 }
@@ -469,8 +468,8 @@ export async function selectWriteableTxtFile(baseName?: string): Promise<FileSys
    return selectWriteableFileImpl(options);
 }
 
-export function getRandom48(): Uint8Array {
-   return sodium.randombytes_buf(48);
+export function getRandom(length: number): Uint8Array<ArrayBuffer> {
+   return ensureArrayBuffer(sodium.randombytes_buf(length));
 }
 
 // Helper function to get bytes from a string, truncated to a maximum byte length, ensuring valid UTF-8